+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
  - Set loop member variable when constructing SpriteAnimationComponent from SpriteAnimationData
  - Fix rendering of children
+ - Properly propagate onMount and onRemove to children
 
 ## 1.0.0-rc3
  - Fix TextBoxComponent rendering